Crate repitile_server

repitile_server uses the repitile_core crate, along with a gotham REST API server to communicate with the UI and update profiles, configurations, temperature, humidity, etc.

Re-exports

extern crate std;

extern crate futures;

extern crate gotham;

extern crate hyper;

extern crate lazy_static;

extern crate dht22_pi;

extern crate mime;

extern crate repitile_core;

extern crate sysfs_gpio;

use std::prelude::v1::*;

use std::thread;

use std::sync::mpsc;

use std::sync::mpsc::Receiver;

use std::sync::mpsc::SyncSender;

use std::sync::Mutex;

use repitile_core::CommReq;

use repitile_core::profile::Profile;

use repitile_core::config::Configuration;

use simple_sensor::SimpleSensor;

use simple_regulator::SimpleHumidityRegulator;

use simple_regulator::SimpleTempRegulator;

Modules

rest_server
Contains the REST API server functionality.

simple_regulator
Defines a simple regulator that enables/disables a LED connected to the Pi for simple threshold testing.

simple_sensor
Defines a simple sensor that reads from our DHT22/RHT03 sensors connected to the Pi.

Structs

GOTHAM_CHANNEL_SERVER_REQ

GOTHAM_CHANNEL_SERVER_RESP

Statics

GOTHAM_CHANNEL_SERVER_REQ

GOTHAM_CHANNEL_SERVER_RESP

Functions

main